+++ mux.go
@@ -1,0 +1,13 @@
+package main
+
+import "net/http"
+
+func NewMux() http.Handler {
+	mux := http.NewServeMux()
+	mux.HandleFunc("/health", func(w http.ResponseWriter, r *http.Request) {
+		w.Header().Set("Content-Type", "application/json; charset=uft-8")
+		// 性的解析のエラーを回避するため明示的に戻り値を捨てている
+		_, _ = w.Write([]byte(`{"status": "ok"}`))
+	})
+	return mux
+}
