+++ server.go
@@ -2,65 +2,48 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"log"
 	"net"
 	"net/http"
 	"os"
 	"os/signal"
 	"syscall"
-	"time"
 
-	"github.com/ky0yk/go_todo_app/config"
 	"golang.org/x/sync/errgroup"
 )
 
-func main() {
-	if err := run(context.Background()); err != nil {
-		log.Printf("failed to terminated server: %v", err)
-		os.Exit(1)
+type Server struct {
+	srv *http.Server
+	l net.Listener
+}
+
+func NewServer(l net.Listener, mux http.Handler) *Server {
+	return &Server{
+		srv: &http.Server{Handler: mux},
+		l: l,
 	}
 }
 
-func run(ctx context.Context) error {
+func (s *Server) Run(ctx context.Context) error {
 	ctx, stop := signal.NotifyContext(ctx, os.Interrupt, syscall.SIGTERM)
 	defer stop()
-	cfg, err := config.New()
-	if err != nil {
-		return err
-	}
-	l, err := net.Listen("tcp", fmt.Sprintf(":%d", cfg.Port))
-	if err != nil {
-		log.Fatalf("failed to listen port %d: %v", cfg.Port, err)
-	}
-	url := fmt.Sprintf("http://%s", l.Addr().String())
-	log.Printf("start with: %v", url)
-	s := &http.Server{
-		// 引数で受け取ったnet.Listenerを利用するので、
-		// Addrフィールドは指定しない
-		Handler: http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-			// コマンドラインで実験するため
-			time.Sleep(5 * time.Second)
-			fmt.Fprintf(w, "Hello, %s!", r.URL.Path[1:])
-		}),
-	}
 	eg, ctx := errgroup.WithContext(ctx)
-	eg.Go(func() error {
-		// ListenAndServeメソッドではなく、Serveメソッドに変更する
-		if err := s.Serve(l); err != nil &&
-			// http.ErrServerClosed は
-			// http.Server.Shutdown() が正常に終了したことを示すので異常ではない。
-			err != http.ErrServerClosed {
+	eg.Go(func() error{
+		// http.ErrServerClosedは
+		// http.Server.Shutdown()が正常に終了したことを示すので異常ではない
+		if err := s.srv.Serve(s.l); err != nil &&
+		err != http.ErrServerClosed {
 			log.Printf("failed to close: %+v", err)
 			return err
 		}
 		return nil
 	})
 
-	<-ctx.Done()
-	if err := s.Shutdown(context.Background()); err != nil {
+	// チャネルから通知（終了通知）を待機する
+	<- ctx.Done()
+	if err := s.srv.Shutdown(context.Background()); err != nil {
 		log.Printf("failed to shutdown: %+v", err)
 	}
-	// グレースフルシャットダウンの終了を待つ。
+	// グレースフルシャットダウンの終了を待つ
 	return eg.Wait()
 }
